refactor: move directoryobject out of util pkg
