+++ pkg/azure/directoryobject/directoryobject.go
