+++ pkg/azure/directoryobject/directoryobject_test.go
